Split openlayers_examples in two modules.
