+++ modules/openlayers_library/src/Plugin/Component/IconSprites/js/iconsprites.js
