+++ modules/openlayers_library/src/Plugin/Component/Permalink/Permalink.php
@@ -1,0 +1,24 @@
+<?php
+/**
+ * @file
+ * Component: Permalink.
+ */
+
+namespace Drupal\openlayers_library\Plugin\Component\Permalink;
+use Drupal\openlayers\Component\Annotation\OpenlayersPlugin;
+use Drupal\openlayers\Openlayers;
+use Drupal\openlayers\Types\Component;
+use Drupal\openlayers\Types\ObjectInterface;
+
+/**
+ * Class Permalink.
+ *
+ * @OpenlayersPlugin(
+ *   id = "Permalink",
+ *   description = "The HTML 5 History API is used to update the browser URL with the current zoom-level, center and rotation when the map is moved."
+ * )
+ *
+ */
+class Permalink extends Component {
+
+}
