+++ modules/openlayers_library/src/Plugin/Component/Permalink/js/permalink.js
