+++ modules/openlayers_library/src/Plugin/Component/ProgressBar/js/progressbar.js
