+++ modules/openlayers_library/src/Plugin/Component/Refresh/Refresh.php
@@ -1,0 +1,21 @@
+<?php
+/**
+ * @file
+ * Component: Refresh.
+ */
+
+namespace Drupal\openlayers_library\Plugin\Component\Refresh;
+use Drupal\openlayers\Component\Annotation\OpenlayersPlugin;
+use Drupal\openlayers\Openlayers;
+use Drupal\openlayers\Types\Component;
+use Drupal\openlayers\Types\ObjectInterface;
+
+/**
+ * Class Refresh.
+ *
+ * @OpenlayersPlugin(
+ *   id = "Refresh"
+ * )
+ *
+ */
+class Refresh extends Component {}
