+++ modules/openlayers_library/src/Plugin/Component/SpinJS/js/spin.js
