+++ modules/openlayers_library/src/Plugin/Component/Swipe/js/swipe.js
