+++ modules/openlayers_library/src/Plugin/Control/OL3LayerSwitcher/js/layerswitcher.js
