+++ modules/openlayers_library/src/Plugin/Control/OL3LayerSwitcher/js/ol3-layerswitcher.js
