+++ modules/openlayers_library/src/Plugin/Style/Timezones/Timezones.php
@@ -1,0 +1,21 @@
+<?php
+/**
+ * @file
+ * Style: Timezones.
+ */
+
+namespace Drupal\openlayers_library\Plugin\Style\Timezones;
+use Drupal\openlayers\Component\Annotation\OpenlayersPlugin;
+use Drupal\openlayers\Openlayers;
+use Drupal\openlayers\Types\ObjectInterface;
+use Drupal\openlayers\Types\Style;
+
+/**
+ * Class Timezones.
+ *
+ * @OpenlayersPlugin(
+ *   id = "Timezones"
+ * )
+ *
+ */
+class Timezones extends Style {}
